refactor(projects): extract SubHireExpandedItems (Phase 7)
Pull the bottom-of-file SubHireExpandedItems helper (~70 LOC + its
own getSubHire query) into its own component. Render is unchanged —
the parent only mounts it when a sub-hire order row in the
Sub-Hire Orders panel is expanded.

equipment-tab.tsx: 1465 → 1393 LOC (-72). Also drops the now-dead
React.Fragment import.

Session final state:
  equipment-tab.tsx: 2148 → 1393 LOC (-755, -35%)
  10 dialogs / helpers extracted to focused files this session
  All 1974 unit + 235 integration tests pass
  Lint: only the long-standing queryKey/useCallback warning remains
