+++ src/components/projects/equipment-tab.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import React, { Fragment, useState, useCallback } from "react";
+import React, { useState, useCallback } from "react";
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import {
   DndContext,
@@ -69,6 +69,7 @@ import { AddCategoryDialog } from "./add-category-dialog";
 import { RenameCategoryDialog } from "./rename-category-dialog";
 import { AddGroupToolbarDialog } from "./add-group-toolbar-dialog";
 import { EditLineItemDialog } from "./edit-line-item-dialog";
+import { SubHireExpandedItems } from "./sub-hire-expanded-items";
 import { SubHireOrderDialog } from "./sub-hire-order-dialog";
 import { getSubHires } from "@/server/sub-hires";
 import { subHireStatusLabels, formatLabel } from "@/lib/status-labels";
@@ -1387,79 +1388,6 @@ export function EquipmentTab({ projectId, rentalStartDate, rentalEndDate }: Equi
           }
         }}
       />
-    </div>
-  );
-}
-
-// ─── Sub-Hire Expanded Items ──────────────────────────────────────────────────
-
-function SubHireExpandedItems({ subHireId, orgId }: { subHireId: string; orgId?: string }) {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const { data: subHire } = useQuery<any>({
-    queryKey: ["sub-hire", orgId, subHireId],
-    queryFn: async () => {
-      const { getSubHire } = await import("@/server/sub-hires");
-      return getSubHire(subHireId);
-    },
-    enabled: !!orgId,
-  });
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const items = (subHire?.items || []) as Array<Record<string, any>>;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const groups = (subHire?.groups || []) as Array<Record<string, any>>;
-  const ungroupedItems = items.filter((item) => !item.groupId);
-
-  if (items.length === 0 && groups.length === 0) {
-    return (
-      <div className="pb-3 ml-4 border-l-2 border-primary/20 pl-8 text-xs text-fg-4 py-2">
-        No items in this order yet.
-      </div>
-    );
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  function renderItemRow(item: Record<string, any>, indent: string) {
-    const itemMargin = Number(item.unitCharge) - Number(item.unitCost);
-    return (
-      <tr key={item.id as string} className="text-sm">
-        <td className={`${indent} py-1.5 text-fg-2`}>
-          {item.description as string}
-          {(item.model as Record<string, string>)?.name && (
-            <span className="ml-1.5 text-xs text-fg-4">({(item.model as Record<string, string>).name})</span>
-          )}
-        </td>
-        <td className="px-3 py-1.5 text-right tabular-nums text-fg-3 w-12">&times;{item.quantity as number}</td>
-        <td className="px-3 py-1.5 text-right tabular-nums text-fg-3 w-24">{formatCurrency(Number(item.unitCost))} cost</td>
-        <td className="px-3 py-1.5 text-right tabular-nums w-24">{formatCurrency(Number(item.unitCharge))}</td>
-        <td className={`px-3 py-1.5 text-right tabular-nums w-20 ${itemMargin > 0 ? "text-success" : itemMargin < 0 ? "text-error" : "text-fg-4"}`}>
-          {formatCurrency(itemMargin)}
-        </td>
-      </tr>
-    );
-  }
-
-  return (
-    <div className="pb-2 ml-4 border-l-2 border-primary/20">
-      <table className="w-full">
-        <tbody>
-          {groups.map((group) => {
-            const groupItems = (group.items || []) as Array<Record<string, unknown>>;
-            return (
-              <Fragment key={group.id}>
-                <tr className="text-xs">
-                  <td colSpan={5} className="pl-8 py-1.5 font-medium text-fg-3">
-                    <span className="text-primary/70">▸</span> {group.title}
-                    <span className="ml-1.5 text-fg-4 font-normal">({groupItems.length} item{groupItems.length !== 1 ? "s" : ""})</span>
-                  </td>
-                </tr>
-                {groupItems.map((item) => renderItemRow(item, "pl-12"))}
-              </Fragment>
-            );
-          })}
-          {ungroupedItems.map((item) => renderItemRow(item, "pl-8"))}
-        </tbody>
-      </table>
     </div>
   );
 }
